requirements.txt
requests
beautifulsoup4
pytesseract
Pillow
PyMuPDF
openai
python-dotenv
pdf2image
tqdm
lxml
regex
